Refactor application into factory package
