+++ harmony/extensions.py
@@ -1,0 +1,15 @@
+from flask_limiter import Limiter
+from flask_limiter.util import get_remote_address
+from flask_migrate import Migrate
+from flask_seasurf import SeaSurf
+from flask_socketio import SocketIO
+from flask_sqlalchemy import SQLAlchemy
+from flask_talisman import Talisman
+
+
+db = SQLAlchemy()
+migrate = Migrate()
+socketio = SocketIO()
+limiter = Limiter(key_func=get_remote_address)
+csrf = SeaSurf()
+talisman = Talisman()
