+++ harmony/models.py
@@ -1,0 +1,38 @@
+from datetime import datetime
+
+from .extensions import db
+
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    spotify_id = db.Column(db.String(100), unique=True, nullable=False)
+    display_name = db.Column(db.String(100), nullable=False)
+    profile_image = db.Column(db.String(200), nullable=True)
+    email = db.Column(db.String(100), nullable=True)
+    top_artists = db.Column(db.JSON, nullable=True)
+    top_tracks = db.Column(db.JSON, nullable=True)
+    genres = db.Column(db.JSON, nullable=True)
+    bio = db.Column(db.Text, nullable=True)
+    age = db.Column(db.Integer, nullable=True)
+    location = db.Column(db.String(255), nullable=True)
+
+
+class Like(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    from_user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    to_user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    created_at = db.Column(db.DateTime, default=db.func.now())
+
+
+class Message(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    sender_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    receiver_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    content = db.Column(db.Text, nullable=False)
+    timestamp = db.Column(db.DateTime, default=datetime.utcnow)
+
+    sender = db.relationship("User", foreign_keys=[sender_id])
+    receiver = db.relationship("User", foreign_keys=[receiver_id])
+
+
+__all__ = ["User", "Like", "Message"]
